test(agents): move parked-consumer watchdog tests to their own file
llm-idle-timeout.test.ts sits at the max-lines lint limit; keep the new
coverage in a sibling file.
